refactor: clearer naming of email lib
