+++ amplify/backend/function/sendNotification/lib/sendEmailNotifications.ts
